+++ source/DangIt/Failure modules/Entropy/Parachutes.cs
@@ -2,11 +2,11 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
-using ippo;
+using nsDangIt;
 using UnityEngine;
 using KSP;
 
-namespace ippo
+namespace nsDangIt
 {
 	public class ModuleParachuteReliability : FailureModule
 	{